Create Two Player
This is an extension of the Single Player version of the Towers of Hanoi game. There are two towers on each side and one in the middle. Each player starts on one side, and must move their tower to one of the towers on the other side. The players can work cooperatively, or not, and can move any piece as long as it is a legal move,. However, only one is the winner.